+++ frontend/src/language/lang/ja.js
@@ -503,6 +503,8 @@ export const m = {
     '知道了': '了解了', 
     '搜索关键字': '検索キーワード', 
     '无资源ID，无法预览': 'リソースIDがないため、プレビューできません', 
+    '暂无可复制的资源实例': 'コピー可能なリソースインスタンスが現在ありません',
+    '暂无可批量复制的资源实例': '一括コピー可能なリソースインスタンスが現在ありません',
     '暂无可批量复制包含有属性条件的资源实例': '属性条件を含むリソースインスタンスの一括コピー可能なものが現在ありません', 
     '(GMT-12:00) 国际换线时间': '(GMT-12:00) 国際日期変更線時間', 
     '(GMT-11:00) 中途岛时间': '(GMT-11:00) ミッドウェー島時間',
