+++ frontend/src/language/lang/zh.js
@@ -510,6 +510,8 @@ export const m = {
         '知道了': '知道了',
         '搜索关键字': '搜索关键字',
         '无资源ID，无法预览': '无资源ID，无法预览',
+        '暂无可复制的资源实例': '暂无可复制的资源实例',
+        '暂无可批量复制的资源实例': '暂无可批量复制的资源实例',
         '暂无可批量复制包含有属性条件的资源实例': '暂无可批量复制包含有属性条件的资源实例',
         '(GMT-12:00) 国际换线时间': '(GMT-12:00) 国际换线时间',
         '(GMT-11:00) 中途岛时间': '(GMT-11:00) 中途岛时间',
