+++ frontend/src/views/grading-admin/components/render-template-resource-instance-table.vue
@@ -55,7 +55,7 @@
                       :key="contentIndex">
                       <div class="content-name">{{ content.name }}</div>
                       <div class="content">
-                        <render-condition
+                        <IamRenderCondition
                           :value="content.value"
                           :is-empty="content.empty"
                           :can-view="item.actions[0].canView"
@@ -106,7 +106,7 @@
                           :key="contentIndex">
                           <div class="content-name">{{ content.name }}</div>
                           <div class="content">
-                            <render-condition
+                            <IamRenderCondition
                               :value="content.value"
                               :is-empty="content.empty"
                               :can-view="actionItem.canView"
@@ -177,12 +177,12 @@
 </template>
 <script>
   import _ from 'lodash';
-  import RenderCondition from '../../perm-apply/components/render-condition';
-  import RenderResource from '../../perm-apply/components/render-resource';
+  import IamRenderCondition from '@/components/iam-render-condition';
+  import RenderResource from '@/views/perm-apply/components/render-resource';
   export default {
     name: '',
     components: {
-      RenderCondition,
+      IamRenderCondition,
       RenderResource
     },
     props: {
