+++ frontend/src/views/manage-spaces/secondary-manage-space/components/render-instance-table.vue
@@ -88,7 +88,7 @@
                   </bk-button>
                 </div>
                 <div class="content">
-                  <render-condition
+                  <IamRenderCondition
                     :ref="`condition_${$index}_aggregateRef`"
                     :value="formatDisplayValue(row)"
                     :is-empty="row.empty"
@@ -119,7 +119,7 @@
                         </template>
                       </div>
                       <div class="content">
-                        <render-condition
+                        <IamRenderCondition
                           data-test-id="group_input_resourceInstanceCondition"
                           :ref="`condition_${$index}_${contentIndex}_ref`"
                           :value="content.value"
@@ -246,8 +246,8 @@
   import RenderAggregateSideslider from '@/components/choose-ip/sideslider';
   import { leaveConfirm } from '@/common/leave-confirm';
   import { CUSTOM_PERM_TEMPLATE_ID, PERMANENT_TIMESTAMP } from '@/common/constants';
+  import IamRenderCondition from '@/components/iam-render-condition';
   import RenderResource from './render-resource';
-  import RenderCondition from './render-condition';
   import PreviewResourceDialog from './preview-resource-dialog';
   import RenderResourcePopover from '@/components/iam-view-resource-popover';
   import RenderDetail from '../common/render-detail';
@@ -260,9 +260,9 @@
       };
     },
     components: {
+      IamRenderCondition,
       RenderAggregateSideslider,
       RenderResource,
-      RenderCondition,
       PreviewResourceDialog,
       RenderResourcePopover,
       RenderDetail
