+++ frontend/src/views/perm-apply/components/resource-instance-table.vue
@@ -52,7 +52,7 @@
               </bk-button>
             </div>
             <div class="group-container">
-              <render-condition
+              <IamRenderCondition
                 :ref="`condition_${$index}_aggregateRef`"
                 :value="formatDisplayValue(row)"
                 :is-empty="row.empty"
@@ -80,7 +80,7 @@
                     </template>
                   </div>
                   <div class="content">
-                    <render-condition
+                    <IamRenderCondition
                       :ref="`condition_${$index}_${contentIndex}_ref`"
                       :value="content.value"
                       :is-empty="content.empty"
@@ -301,8 +301,8 @@
   import AggregationPolicy from '@/model/aggregation-policy';
   import { leaveConfirm } from '@/common/leave-confirm';
   import { PERMANENT_TIMESTAMP } from '@/common/constants';
+  import IamRenderCondition from '@/components/iam-render-condition';
   import RenderResource from './render-resource';
-  import RenderCondition from './render-condition';
   import EffectTime from './effect-time';
   import SidesliderEffectTime from './sideslider-effect-time';
   import PreviewResourceDialog from './preview-resource-dialog';
@@ -324,7 +324,7 @@
     components: {
       RenderAggregateSideslider,
       RenderResource,
-      RenderCondition,
+      IamRenderCondition,
       PreviewResourceDialog,
       EffectTime,
       SidesliderEffectTime
