+++ frontend/src/views/perm-template/edit/sync.vue
@@ -35,7 +35,7 @@
                     :key="`${index}${contentIndex}`">
                     <div class="content-name">{{ content.name }}</div>
                     <div class="content">
-                      <render-condition
+                      <IamRenderCondition
                         :ref="`condition_${index}_${$index}_${contentIndex}_ref`"
                         :value="content.value"
                         :params="curCopyParams"
@@ -192,7 +192,7 @@
   import SyncPolicy from '@/model/template-sync-policy';
   import Condition from '@/model/condition';
   import RenderResourcePopover from '@/components/iam-view-resource-popover';
-  import RenderCondition from '../components/render-condition';
+  import IamRenderCondition from '@/components/iam-render-condition';
   import RenderDetail from '../components/render-detail';
   import RenderResource from '../components/render-resource';
   export default {
@@ -203,8 +203,8 @@
       };
     },
     components: {
+      IamRenderCondition,
       RenderResourcePopover,
-      RenderCondition,
       RenderDetail,
       RenderResource
     },
